BUG: Fix off-by-one bug in end frame for 3D registration
The UI slider in the 3DHierarchicalRegistration module for the range of frames
to be registered implies that the selected interval is inclusive of the end
frame, which is evident also in the number of frames in the initialized
transform sequence. However, the registration used to be run on all frames
except the last one, so the last frame in the resulting transform sequence was
always left as the initialized identity.

This commit fixes this off-by-one bug in the frames being registered.