BUG: Fix generated partial volumes misaligned with input volume node (#147)
This commit fixes how the generated PV tiffs are loaded, to ensure that
the partial volume nodes preserve the original RAS/LPS orientation of
the input volume node.

Previously, the generated tiff files were always loaded in LPS-like
orientation (e.g., the IJK-to-RAS Direction Matrix would have -1 in the
first two diagonal entries). This would work correctly on most data we
have tested against, when the data is also originally in LPS
orientation. However, for data this is inherently in RAS directions
(e.g., the IJK-to-RAS Direction Matrix would have 1 in the first two
diagonal entries), or for cropped volume nodes (cropping is observed to
flip the volume from RAS to LPS directions or vice-versa), the loaded
PVs would appear misaligned.

Note: Since TIFF files don't store the directions of the volume node,
volume files opened in Slicer through the "Add Data" widget are not 
guaranteed to be loaded with the right IJK-to-RAS directions. For this 
reason, the generated volumes must be imported using the "Load Partial 
Volumes" button under the "Partial Volume Generation" group when the 
corresponding input volume node is selected in the Pre-Processing 
module.