ENH: Improve pre-processing error handling for UI user friendliness (#153)
As a follow up to #141, this PR improves the error reporting of the
pre-processing module for better user friendliness. Namely, the error
handling code was refactored to better report a meaningful error
message to the UI user when an exception is raised (like what
specific field(s) has the invalid value). The error is still reported
to the error log for CLI users.

Co-authored-by: Jean-Christophe Fillion-Robin <jchris.fillionr@kitware.com>